requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
Django==4.1.5
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
exceptiongroup==1.1.0
iniconfig==1.1.1
packaging==22.0
pluggy==1.0.0
pytest==7.2.0
pytest-django==4.5.2
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
